refactor: split agent-fleet into sections and pure scroll modules
AgentFleet.tsx was 830 lines: six page sections, three scroll-driven effects,
and a component defined inside the render body.

Pure math moves to lib/agent-system: links.ts holds the bezier geometry,
belt.ts the scrub progress and nearest-station argmin, reveal.ts the count-up
easing and scene clamp, clipboard.ts the copy helper. The hardcoded /9 at
three sites becomes STATION_COUNT derived from STATIONS.length, and the
breakpoint, debounce, duration and threshold literals are named constants.

The three effects become useMapLinks, useBeltScrub and useReveal in hooks.ts,
each keeping its own rAF flag and listener teardown. Markup splits into six
section components plus Glyph, Machine, Divider, Station and NodeButton.
Hero coords, divider paths, legend rows, team lists and the save-job prompt
move to data/agent-fleet.ts as additive exports.

NodeButton is hoisted from the render body to module scope. This is the one
non-behavior-preserving change: it previously got a fresh component identity
every render, so React unmounted and remounted it, restarting focus and
transitions. It now takes name, roleClass, selected, onSelect and onHover.

State follows its markup - route into IntakeSection, selected and litName into
FleetMapSection - leaving AgentFleet.tsx at 29 lines with no prop drilling.
